+++ crates/web-client/src/models/accounts.rs
@@ -3,7 +3,7 @@ use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
 #[derive(Serialize, Deserialize)]
-pub struct SerializedAccountStub {
+pub struct SerializedAccountHeader {
     id: String,
     nonce: String,
     vault_root: String,
@@ -12,15 +12,15 @@ pub struct SerializedAccountStub {
 }
 
 #[wasm_bindgen]
-impl SerializedAccountStub {
+impl SerializedAccountHeader {
     pub fn new(
         id: String,
         nonce: String,
         vault_root: String,
         storage_root: String,
         code_root: String,
-    ) -> SerializedAccountStub {
-        SerializedAccountStub {
+    ) -> SerializedAccountHeader {
+        SerializedAccountHeader {
             id,
             nonce,
             vault_root,