+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.6.0 (TBD)
 
+* [BREAKING] Adds the `TransactionProver` trait to the `Client` struct (#517).
+- [BREAKING] Renamed `off-chain` and `on-chain` to `private` and `public` respectively for the account storage modes (#516).
 * [BREAKING] Re-exported `TransactionRequest` from submodule, renamed `AccountDetails::Offchain` to `AccountDetails::Private`, renamed `NoteDetails::OffChain` to `NoteDetails::Private` (#508).
 
 ## v0.5.0 (2024-08-27)